Separate cypress commands into modules
