+++ cypress/support/commands/index.ts
@@ -1,0 +1,5 @@
+/// <reference types="cypress" />
+import "@testing-library/cypress/add-commands";
+import "cypress-network-idle";
+import "./select";
+import "./image";
